backend/requirements.txt
fastapi>=0.104.0
uvicorn[standard]>=0.24.0
websockets>=12.0
python-socketio>=5.10.0
pydantic>=2.0.0
python-multipart>=0.0.6
sqlalchemy>=2.0.0
python-dotenv>=1.0.0
pydantic-settings>=2.0.0
